Add badges invertable option.
Add more request specs for admin part.

Adjust gem specs to recent badges structure changes.

+++ app/models/concerns/ff_badges/user.rb
@@ -40,7 +40,7 @@ module Concerns::FfBadges::User
   end
 
   def remove_badge!(badge)
-    if self.has_badge?(badge)
+    if self.has_badge?(badge) && badge.invertable?
       self.user_badges.where(badge_id: badge.id).first.destroy
     end
   end
